Bug fixes for windows
cmaes.cpp: abs is overloaded to accept either double or float, here cast it to
double to resolve the ambiguity

cma_es_ext.cpp: see here:
http://msdn.microsoft.com/en-us/library/ms173598(v=vs.80).aspx

+++ cma_es/cma/cmaes.cpp
@@ -2270,7 +2270,7 @@ random_init( random_t *t, long unsigned inseed)
   if (inseed < 1) {
     while ((long) (cloc - clock()) == 0)
       ; /* TODO: remove this for time critical applications? */
-    inseed = (long unsigned)abs(100*time(NULL)+clock());
+    inseed = (long unsigned)abs((double)(100*time(NULL)+clock()));
   }
   return random_Start(t, inseed);
 }
@@ -2633,7 +2633,7 @@ readpara_SupplementDefaults(readpara_t *t)
   if (t->seed < 1) {
     while ((int) (cloc - clock()) == 0)
       ; /* TODO: remove this for time critical applications!? */
-    t->seed = (unsigned int)abs(100*time(NULL)+clock());
+    t->seed = (unsigned int)abs((double)(100*time(NULL)+clock()));
   }
 
   if (t->stStopFitness.flg == -1)
